refactor: extract shared DevSignInButton UI + dev credentials
Replaces apps/desktop's LocalDevAuthForm (email/password form) with a
one-click button matching the web. Both apps now wrap a single shared
UI component and supply only platform-specific glue:

  packages/ui/.../dev-sign-in-button.tsx   pure UI button + state
  packages/shared/dev-credentials.ts       DEV_EMAIL/PASSWORD/NAME

  apps/web/.../DevSignInButton.tsx         Next.js useRouter +
                                           authClient.signIn.email
  apps/desktop/.../LocalDevSignInButton/   TanStack useNavigate +
                                           raw fetch + electron-trpc
                                           token persistence

Eliminates dev-cred duplication across four files (db seed, desktop
main auto-sign-in, desktop renderer wrapper, web wrapper). The
desktop sign-in still uses raw fetch + saveToken because Electron
manages tokens itself; the web wrapper uses cookies via
@superset/auth/client. Routing diverges naturally between the two
routers, which is why the shared piece is just the button itself.

Dev-only on both platforms — desktop gates on SUPERSET_PROFILE=local,
web gates on NODE_ENV !== "production".
