+++ apps/desktop/src/renderer/routes/sign-in/components/LocalDevSignInButton/index.ts
@@ -1,0 +1,1 @@
+export { LocalDevSignInButton } from "./LocalDevSignInButton";
