+++ packages/shared/package.json
@@ -12,6 +12,10 @@
 			"types": "./src/deployment-profile.ts",
 			"default": "./src/deployment-profile.ts"
 		},
+		"./dev-credentials": {
+			"types": "./src/dev-credentials.ts",
+			"default": "./src/dev-credentials.ts"
+		},
 		"./v2-only-user": {
 			"types": "./src/v2-only-user.ts",
 			"default": "./src/v2-only-user.ts"
